fix segfault
recv was segfaulting, debugging this it turns out this was because of
c_nn_freemsg call. I commented it, could it be that aloca clears the
pointer correctly already? I ran a server with limited heap in
Haskell, flooded it with messages and it seems the heap didn't
increase in size.